Softomation/HighwaySolutions/WebApi/TMSPyAPI/requirements.txt
Flask
PyJWT
paho-mqtt
cryptography
pyodbc